+++ Modules/FSRMDsc/DSCResources/DSR_FSRMFileScreenTemplateAction/DSR_FSRMFileScreenTemplateAction.psm1
@@ -90,7 +90,7 @@ function Get-TargetResource
     {
         Write-Verbose -Message ( @(
                 "$($MyInvocation.MyCommand): "
-                $($LocalizedData.ActionDoesNotExistMessage) `
+                $($LocalizedData.ActionNotExistMessage) `
                     -f $Name, $Type
             ) -join '' )
 
@@ -676,8 +676,18 @@ function Test-TargetResource
                 $desiredConfigurationMatch = $false
             } # if
 
-            if (($PSBoundParameters.ContainsKey('ReportTypes')) `
-                    -and ($action.ReportTypes -ne $ReportTypes))
+            # Get the existing report types into an array
+            if ($null -eq $action.ReportTypes)
+            {
+                [System.String[]] $existingReportTypes = @()
+            }
+            else
+            {
+                [System.String[]] $existingReportTypes = $action.ReportTypes
+            }
+
+            if ($PSBoundParameters.ContainsKey('ReportTypes') -and `
+                (Compare-Object -ReferenceObject $existingReportTypes -DifferenceObject $ReportTypes).Count -ne 0)
             {
                 Write-Verbose -Message ( @(
                         "$($MyInvocation.MyCommand): "